adding components for webpage
